Add Storage abstraction over asyncStorage (#1074)
* Add Storage abstraction over asyncStorage

- Adds a Storage context provider which allows for sync and reactive access to all storage values
- For the start, only wordNodeCards are ported from the old solution to the storage context

* Make write async

The most important thing is that read is sync, but write can be async without too much loss of ergonomics.
Also, it was probably incorrect before.

* Port `isTrackingEnabled`

* Port `selectedProfessions`

* Add 'StorageCache' and 'useStorage'

This is mostly an optimization. Before, when the `StorageContext` changed, every component that used it would rerender. This solution is a lot more granular and only components that use the changed value will rerender.

* Port `isDevModeEnabled`

* port `progress`

* Port `cmsUrlOverwrite`

* Port `customDisciplines`

* Storage: Make it clear that `getItem` returns an immutable value

- StorageCache always returns the same object. If somebody modifies it, it will change in subsequent calls even though the storage has not changed.
- React requires the value returned by `getItem` to be the same object every time though. Otherwise, `useSyncExternalStore` would cause an inifinite loop: https://react.dev/reference/react/useSyncExternalStore#im-getting-an-error-the-result-of-getsnapshot-should-be-cached

As solution, I decide to add `getMutableItem` to the storage cache which makes a deep copy and mark the return type of `getItem` as immutable. Hopefully this will catch all invalid uses of `getItem`.

* Port `userVocabulary`

* Port `nextUserVocabularyId`

* Port `favorites`

- Instead of migrating the storage format on every access,
  only do it once at startup
- This allows to continue using `storageCache.getItem('favorites')`

* Rename `AsyncStorage` to `storageUtils`

The old name did not really fit anymore, because the storage functionality has moved and async storage only contained utility functions.

Also, ideally, we should probably extract the functionality into different services and get rid of this file. This has low priority though.

* Add Storage test

* Apply suggestions

* Rename `StorageCache.createForTesting` to `createDummy`

* Replace `getStorageKey` with an object

* Add `useStorageCache` hook