Fix crash on scope selection
Before, when a profession was selected via search, the app
attempted to navigate to `ManageSelection`.
This was only valid from the `Lernbereiche verwalten` screen however,
and not possible from the first start screen.

The new behavior in this commit is to not navigate anywhere and
enable toggling the list items.