Enable tts for ios
Tts not working on ios was a bug actually. It did not work, because the
tts engine returns `true` instead of `'success'` for some reason.
I think this fix to just remove the check should work, because
this is how it is suggested in the repo: https://github.com/ak1394/react-native-tts?tab=readme-ov-file#waiting-for-initialization