Debug modal improvements (#1093)
* Add dev setting to clear selected professions

This makes it a bit easier to test the initial screen,
especially on ios where I could not find a quick way to clear app
data, except programmatically.

* Add open debug modal button to settings

This is only shown if dev mode is enabled.
The purpose of this is to make it slightly easier to access the
dev settings.