Use app toolbelt (#1128)
* Integrate the App-toolbelt

This changes how releases are generated. Before, a new release generated two new tags, one for android and one for ios.
Now we create just one release for both and remove the option of only releasing for a single platform. This should simplify the release process.
Additionally, it is not possible anymore to create a production release without a beta release before, so the scripts related to that functionality have also been removed.

Changes:
- Depend on the app-toolbelt
- Get rid of the old tools
- unify notify_android and notify_ios
- Create translations/store-translations/lunes.json with some dummy values for the app toolbelt
- Add .github/release.yml so that github can generate release notes when new releases are created