Various CI fixes and improvements (#1131)
* Fix mattermost link to the release notes

* Specify the app name on promotion message

Otherwise there could be confusion about the project

* Include apk and ipa files in mattermost notification for new release

Due to an oversight this was not the case already.
The apk-url and ipa-files have to exist for this feature to work.

* Set correct number of available CPUs

The resource class `xlarge` has 8 total CPUs:
https://circleci.com/pricing/price-list/

* Specify `small` resource class for scripts that don't need more

Where unspecified, circleci defaults to `medium`, which costs twice as
much and is not required for these simple scripts