Implement a basic mechanism for migrations (#1175)
* Implement a basic mechanism for migrations

For now, we only use two versions:
- 0, which represents every version where no version information is stored yet and
- 1 for the current storage version

Now the migration to the new favorites format only happens once, on upgrade from version 0 to version 1

* Run migrations before `storageCache` gets constructed

Otherwise, the migrations will not be maintainable.
For example, the favorites migration used `storageCache.setItem` to store the favorites in the new format.
However, if we decide to change the favorite representation again in the future, this code will become invalid.