Add back manually triggerable workflow to upload a build to browserstack
This mostly reverts commit 9eb25ef6a46460732a7eb2637e2bd3691e070561

The only difference is that the upload to browserstack does not get triggered automatically anymore.

The reason for reverting the commit is that we want to make it simpler for the ui/ux team to test pull requests.
This is a first step towards the goal, because it allows testing a pull request in the web without having a dev environment configured.